add option to estimate amount of meals from existing ingredients
The mod apparently assumes that all raw food is immediately cooked
into meals and works poorly if meals are cooked on demand: If preferability
is set to meals, it gives only the amount of days for which meals are
cooked on demand. If raw food is counted too, it gives an estimate that is
too low because cooking meals usually increases nutrition value.

Add an option (enabled by default) that tries to convert cooking ingredients
into configured meal option.